Ajustando funcao logic->checkin para gerar senha. Atualização nas telas para exibir a senha gerada.
